Use smaller integer types for pointers
Use u32 as the pointer size for arena::P, arena::S, ValueRef and StmtId.
Checks were added so if the representation is not sufficient in the
production, it will crash and we can get notified. This reduced the size
of TraceItem from 24 bytes to 8 bytes, which should help to fit it into
memory.

+++ core/src/data/parser.rs
@@ -370,9 +370,7 @@ impl<'a, 'b, R: BufRead> Parser<'a, 'b, R> {
     fn parse_stmt_id(&mut self) -> ParseResult<StmtId> {
         let file_id = self.parse_file_id()?;
         let stmt_id = extend_buffer!(self, self.parse_u64()?);
-        Ok(StmtId::new(
-            self.arenas.stmt_id.get((file_id, stmt_id)) as u64
-        ))
+        Ok(StmtId::new(self.arenas.stmt_id.get((file_id, stmt_id))))
     }
 
     fn parse_file_id(&mut self) -> ParseResult<FileId> {
